+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 17.07.2018
 
+* [Rust] Code of Kutulu - Moving to the closest explorer (CodinGame)
 * [Rust] Code of Kutulu - Closest explorer with a HashMap (CodinGame)
 
 #### 16.07.2018